requirements.txt
taipy
supabase
pandas
python-pptx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
